Use ` as quote for columns in formula
In order to have the same syntax between `query` postprocess
and `formula` postprocess, let's escape column names with '`'.
The support of the old syntax is done but with a deprecation message.
There is also a method `get_new_syntax_formula` that converts the
old syntax into the new one
(Could be useful if a migration on the front config was planned)